Passenger add

+++ src/main/java/uz/ccrew/flightmanagement/dto/user/UserDTO.java
@@ -5,5 +5,5 @@ import uz.ccrew.flightmanagement.enums.UserRole;
 import lombok.Builder;
 
 @Builder
-public record UserDTO(Integer id, String login, UserRole role) {
+public record UserDTO(Long id, String login, UserRole role) {
 }
